Adapt DrQ-v2 for modern env & add n-step ablation study
- Update for Python 3.12 + mujoco 3.x + dm_control 1.x
- Upgrade hydra-core to 1.3.2 (version_base='1.1' for compat)
- Fix float64/float32 dtype mismatch with newer dm_control
- Fix replay_buffer worker seed for Python 3.12
- Add Windows support (skip MUJOCO_GL=egl)
- Add scripts/run_ablation.sh for n-step ablation experiments
- Update README with reproduction info and modern setup instructions

+++ drqv2.py
@@ -237,9 +237,12 @@ class DrQV2Agent:
         obs, action, reward, discount, next_obs = utils.to_torch(
             batch, self.device)
 
-        # augment
+        # augment and ensure float32
         obs = self.aug(obs.float())
         next_obs = self.aug(next_obs.float())
+        action = action.float()
+        reward = reward.float()
+        discount = discount.float()
         # encode
         obs = self.encoder(obs)
         with torch.no_grad():
